+++ Tests/FluentAssertions.Specs/Primitives/StringAssertionSpecs.cs
@@ -3247,40 +3247,53 @@ namespace FluentAssertions.Specs
         #region (Not) Upper
 
         [Fact]
-        public void Should_succeed_when_asserting_upper_string_to_be_upper()
+        public void When_an_upper_case_string_is_supposed_to_be_in_upper_case_only_it_should_not_throw()
         {
             // Arrange
             string actual = "ABC";
 
             // Act / Assert
-            actual.Should().BeUpper();
+            actual.Should().BeUpperCased();
         }
 
         [Fact]
-        public void Should_fail_when_asserting_non_upper_string_to_be_upper()
+        public void When_a_non_upper_string_is_supposed_to_be_upper_it_should_throw()
         {
             // Arrange
             string actual = "abc";
 
             // Act
-            Action act = () => actual.Should().BeUpper();
+            Action act = () => actual.Should().BeUpperCased();
 
             // Assert
             act.Should().Throw<XunitException>();
         }
 
         [Fact]
-        public void Should_fail_with_descriptive_message_when_asserting_non_upper_string_to_be_upper()
+        public void When_an_upper_case_string_with_numbers_is_supposed_to_be_in_upper_case_only_it_should_throw()
+        {
+            // Arrange
+            string actual = "A1";
+
+            // Act
+            Action act = () => actual.Should().BeUpperCased();
+
+            // Assert
+            act.Should().Throw<XunitException>();
+        }
+
+        [Fact]
+        public void When_a_non_upper_string_is_supposed_to_be_upper_it_should_fail_with_descriptive_message()
         {
             // Arrange
             string actual = "abc";
 
             // Act
-            Action act = () => actual.Should().BeUpper("because we want to test the failure {0}", "message");
+            Action act = () => actual.Should().BeUpperCased("because we want to test the failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                @"Expected actual to be upper because we want to test the failure message, but found ""abc"".");
+                @"Expected all characters in actual to be upper cased because we want to test the failure message, but found ""abc"".");
         }
 
         [Fact]
@@ -3290,58 +3303,68 @@ namespace FluentAssertions.Specs
             string nullString = null;
 
             // Act
-            Action act = () => nullString.Should().BeUpper("because strings should never be {0}", "null");
+            Action act = () => nullString.Should().BeUpperCased("because strings should never be {0}", "null");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Expected nullString to be upper because strings should never be null, but found <null>.");
+                "Expected all characters in nullString to be upper cased because strings should never be null, but found <null>.");
         }
 
         [Fact]
-        public void Should_succeed_when_asserting_non_upper_string_to_be_non_upper()
+        public void When_a_non_upper_string_is_supposed_to_be_non_upper_it_should_succeed()
         {
             // Arrange
             string actual = "abc";
 
             // Act / Assert
-            actual.Should().NotBeUpper();
+            actual.Should().NotBeUpperCased();
         }
 
         [Fact]
-        public void When_asserting_null_string_to_not_be_upper_it_should_succeed()
+        public void When_a_null_string_is_not_supposed_to_be_upper_it_should_succeed()
         {
             // Arrange
             string actual = null;
 
             // Act / Assert
-            actual.Should().NotBeUpper();
+            actual.Should().NotBeUpperCased();
         }
 
         [Fact]
-        public void Should_fail_when_asserting_upper_string_not_to_be_upper()
+        public void When_an_upper_string_is_not_supposed_to_be_upper_it_should_throw()
         {
             // Arrange
             string actual = "ABC";
 
             // Act
-            Action act = () => actual.Should().NotBeUpper();
+            Action act = () => actual.Should().NotBeUpperCased();
 
             // Assert
             act.Should().Throw<XunitException>();
         }
 
         [Fact]
-        public void Should_fail_with_descriptive_message_when_asserting_upper_string_not_to_be_upper()
+        public void When_an_upper_case_string_with_numbers_is_not_supposed_to_be_in_upper_case_only_it_should_succeed()
+        {
+            // Arrange
+            string actual = "a1";
+
+            // Act / Assert
+            Action act = () => actual.Should().NotBeUpperCased();
+        }
+
+        [Fact]
+        public void When_an_upper_string_is_not_supposed_to_be_upper_it_should_fail_with_descriptive_message()
         {
             // Arrange
             string actual = "ABC";
 
             // Act
-            Action act = () => actual.Should().NotBeUpper("because we want to test the failure {0}", "message");
+            Action act = () => actual.Should().NotBeUpperCased("because we want to test the failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Did not expect actual to be upper because we want to test the failure message.");
+                "Did not expect any characters in actual to be upper cased because we want to test the failure message.");
         }
 
         #endregion (Not) Upper
@@ -3349,50 +3372,63 @@ namespace FluentAssertions.Specs
         #region (Not) Lower
 
         [Fact]
-        public void Should_succeed_when_asserting_lower_string_to_be_lower()
+        public void When_a_lower_string_is_supposed_to_be_lower_it_should_succeed()
         {
             // Arrange
             string actual = "abc";
 
             // Act / Assert
-            actual.Should().BeLower();
+            actual.Should().BeLowerCased();
         }
 
         [Fact]
-        public void Should_succeed_when_asserting_empty_string_to_be_lower()
+        public void When_an_empty_string_is_supposed_to_be_lower_it_should_succeed()
         {
             // Arrange
             string actual = "";
 
             // Act / Assert
-            actual.Should().BeLower();
+            actual.Should().BeLowerCased();
         }
 
         [Fact]
-        public void Should_fail_when_asserting_non_lower_string_to_be_lower()
+        public void When_a_non_lower_string_is_supposed_to_be_lower_it_should_fail()
         {
             // Arrange
             string actual = "ABC";
 
             // Act
-            Action act = () => actual.Should().BeLower();
+            Action act = () => actual.Should().BeLowerCased();
 
             // Assert
             act.Should().Throw<XunitException>();
         }
 
         [Fact]
-        public void Should_fail_with_descriptive_message_when_asserting_non_lower_string_to_be_lower()
+        public void When_a_lower_case_string_with_numbers_is_supposed_to_be_in_lower_case_only_it_should_throw()
+        {
+            // Arrange
+            string actual = "a1";
+
+            // Act
+            Action act = () => actual.Should().BeLowerCased();
+
+            // Assert
+            act.Should().Throw<XunitException>();
+        }
+
+        [Fact]
+        public void When_a_non_lower_string_is_supposed_to_be_lower_it_should_fail_with_descriptive_message()
         {
             // Arrange
             string actual = "ABC";
 
             // Act
-            Action act = () => actual.Should().BeLower("because we want to test the failure {0}", "message");
+            Action act = () => actual.Should().BeLowerCased("because we want to test the failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Expected actual to be lower because we want to test the failure message, but found \"ABC\".");
+                "Expected all characters in actual to be lower cased because we want to test the failure message, but found \"ABC\".");
         }
 
         [Fact]
@@ -3402,58 +3438,68 @@ namespace FluentAssertions.Specs
             string nullString = null;
 
             // Act
-            Action act = () => nullString.Should().BeLower("because strings should never be {0}", "null");
+            Action act = () => nullString.Should().BeLowerCased("because strings should never be {0}", "null");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Expected nullString to be lower because strings should never be null, but found <null>.");
+                "Expected all characters in nullString to be lower cased because strings should never be null, but found <null>.");
         }
 
         [Fact]
-        public void Should_succeed_when_asserting_non_lower_string_to_be_upper()
+        public void When_a_non_lower_string_is_supposed_to_be_upper_it_should_succeed()
         {
             // Arrange
             string actual = "ABC";
 
             // Act / Assert
-            actual.Should().NotBeLower();
+            actual.Should().NotBeLowerCased();
         }
 
         [Fact]
-        public void When_asserting_null_string_to_not_be_lower_it_should_succeed()
+        public void When_a_null_string_is_not_supposed_to_be_lower_it_should_succeed()
         {
             // Arrange
             string actual = null;
 
             // Act / Assert
-            actual.Should().NotBeLower();
+            actual.Should().NotBeLowerCased();
         }
 
         [Fact]
-        public void Should_fail_when_asserting_lower_string_to_be_upper()
+        public void When_a_lower_string_is_supposed_to_be_upper_it_should_throw()
         {
             // Arrange
             string actual = "abc";
 
             // Act
-            Action act = () => actual.Should().NotBeLower();
+            Action act = () => actual.Should().NotBeLowerCased();
 
             // Assert
             act.Should().Throw<XunitException>();
         }
 
         [Fact]
-        public void Should_fail_with_descriptive_message_when_asserting_lower_string_to_be_not_lower()
+        public void When_a_lower_case_string_with_numbers_is_not_supposed_to_be_in_lower_case_only_it_should_succeed()
+        {
+            // Arrange
+            string actual = "A1";
+
+            // Act / Assert
+            Action act = () => actual.Should().NotBeLowerCased();
+        }
+
+        [Fact]
+        public void When_a_lower_string_is_not_supposed_to_be_lower_it_should_fail_with_descriptive_message()
         {
             // Arrange
             string actual = "abc";
 
             // Act
-            Action act = () => actual.Should().NotBeLower("because we want to test the failure {0}", "message");
+            Action act = () => actual.Should().NotBeLowerCased("because we want to test the failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>().WithMessage(
-                "Did not expect actual to be lower because we want to test the failure message.");
+                "Did not expect any characters in actual to be lower cased because we want to test the failure message.");
         }
 
         #endregion (Not) Lower